requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
Flask==2.2.3
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
pycparser==2.21
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
